% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset201_ushcn_chihuahuan.R
\name{derive201}
\alias{derive201}
\title{Create (and write) dataset 201 - climate metrics for 9 Chihuahuan desert
USHCN sites}
\usage{
derive201(fname = "ushcn_chihuahuan_derived.csv", dest_path = NULL)
}
\arguments{
\item{fname}{Name of the file to write if dest_path != NULL}

\item{dest_path}{Path to write fname to. No file written if NULL (default)}
}
\value{
A dataframe with derived data
}
\description{
Create (and write) dataset 201 - climate metrics for 9 Chihuahuan desert
USHCN sites
}
