% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset103_ushcn_chihuahuan.R
\name{dev_save_103}
\alias{dev_save_103}
\title{Update the Chihuahuan Desert USHCN DERIVED dataset}
\usage{
dev_save_103(new_derived)
}
\arguments{
\item{new_derived}{New dataset 103 to replace \code{ushcn_chihuahuan_derived.rda}}
}
\value{
A new rda file in data/
}
\description{
Update the package dataset \code{data/ushcn_chihuahuan_derived.rda}. This is not
exported so use \code{devtools::load_all()} to access.
}
