% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset202_jornada_npp.R
\name{dev_save_202}
\alias{dev_save_202}
\title{Update the Jornada NPP DERIVED dataset}
\usage{
dev_save_202(new_derived)
}
\arguments{
\item{new_derived}{New dataset 202 to replace \code{jrn_npp_derived.rda}}
}
\value{
A new rda file in 'data/'
}
\description{
Update the package dataset \code{data/jrn_NPP_derived.rda}. This is not
exported so use \code{devtools::load_all()} to access.
}
