% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{from_edi}
\alias{from_edi}
\title{Get revisions and load latest data entity from an EDI dataset}
\usage{
from_edi(pkgid, rrev = 0, scope = "knb-lter-jrn", skip = 0)
}
\arguments{
\item{pkgid}{An integer defining the dataset ID to retrieve from EDI}

\item{rrev}{A zero (default) or negative integer indicating dataset revision
relative to the most current revision (-1 is the previous)}

\item{scope}{= The EDI scope (string) to retrieve the dataset from}

\item{skip}{= The number of lines to skip on reading the}
}
\value{
A dataframe loaded from EDI
}
\description{
Get revisions and load latest data entity from an EDI dataset
}
