% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jrn_npp_derived}
\alias{jrn_npp_derived}
\title{Dataset 104 - Monthly derived climate metrics from the 15 Jornada LTER NPP
weather stations.}
\format{
A data frame with 1513 rows and 15 variables:
\describe{
\item{year}{Year of observation, integer}
\item{month}{Month of observation, integer}
\item{sitename}{NPP site identifier code, 4-letter string}
...
}
}
\source{
\describe{Derived from published EDI datasets and the
derive104_jornada_npp.R file}
}
\usage{
jrn_npp_derived
}
\description{
A dataset containing stationids, names, location information, and time series
of met and derived climate indices (SPEI, scPDSI) for the 15 weather
stations in the Jornada Basin (southern New Mexico, U.S.A.) that are
associated with the long-term NPP study. Published to EDI as Jornada dataset
104.
}
\keyword{datasets}
