% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_chihuahuan_USHCN}
\alias{update_chihuahuan_USHCN}
\title{Update the Chihuahuan Desert USHCN dataset}
\usage{
update_chihuahuan_USHCN(
  ushcn_path = "/home/greg/data/rawdata/NCDC/ushcn_v2.5/ushcn.v2.5.5.20220609/",
  condaenv_path = "/home/greg/data/miniconda3/bin/conda",
  dest_path = NULL
)
}
\arguments{
\item{ushcn_path}{Path to a downloaded archive of the USHCN dataset}

\item{condaenv_path}{Path to the climatederiv conda environment on the host}

\item{dest_path}{Path for output csv file (default NULL, with no output csv)}
}
\value{
A dataframe loaded from EDI
}
\description{
Update the Chihuahuan Desert USHCN dataset
}
