% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushcn_chihuahuan_data}
\alias{ushcn_chihuahuan_data}
\title{Long-term met data from 9 Chihuahuan desert USHCN weather stations}
\format{
A data frame with 28776 rows and 12 variables:
\describe{
\item{stationid}{USHCN station identifier, unique alphanumeric}
\item{date}{Date of observation, POSIXct}
\item{year}{Year of observation, integer}
...
}
}
\source{
\url{https://www.ncei.noaa.gov/products/land-based-station/us-historical-climatology-network}
}
\usage{
ushcn_chihuahuan_data
}
\description{
A dataset containing stationids, names, location information, and time series
of temperature and precipitation data from the 9 USHCN stations located in
the northern Chihuahuan Desert (U.S.A. portion).
}
\keyword{datasets}
