% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushcn_chihuahuan_derived}
\alias{ushcn_chihuahuan_derived}
\title{Dataset 103 - Monthly derived climate metrics from 9 Chihuahuan desert USHCN
weather stations.}
\format{
A data frame with 11988 rows and 15 variables:
\describe{
\item{stationid}{USHCN station identifier, unique alphanumeric}
\item{date}{Date of observation, POSIXct}
\item{year}{Year of observation, integer}
...
}
}
\source{
\describe{Derived from the ushcn_chihuahuan_data dataset and
derive103_ushcn_chihuahuan.R file}
}
\usage{
ushcn_chihuahuan_derived
}
\description{
A dataset containing stationids, names, location information, and time series
of met and derived climate indices (SPEI, scPDSI) for the 9 USHCN stations
located in the northern Chihuahuan Desert (U.S.A. portion).  Published to
EDI as Jornada dataset 103.
}
\keyword{datasets}
