% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushcn_chihuahuan_stations}
\alias{ushcn_chihuahuan_stations}
\title{Station descriptors for 9 Chihuahuan desert USHCN weather stations}
\format{
A data frame with 9 rows and 6 variables:
\describe{
\item{stationid}{USHCN station identifier, unique alphanumeric}
\item{lat}{Latitude of station, decimal degrees}
\item{lon}{Longitude of station, decimal degrees}
...
}
}
\source{
\url{https://www.ncei.noaa.gov/products/land-based-station/us-historical-climatology-network}
}
\usage{
ushcn_chihuahuan_stations
}
\description{
Table of station id, names, and location information for the 9 USHCN stations
located in the northern Chihuahuan Desert (U.S.A. portion).
}
\keyword{datasets}
