% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset103_ushcn_chihuahuan.R
\name{plot_103}
\alias{plot_103}
\title{Make some plots of the 9-site Chihuahuan Desert USHCN dataset}
\usage{
plot_103()
}
\value{
Some figures are printed
}
\description{
Make some plots of the 9-site Chihuahuan Desert USHCN dataset
}
