% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushcn_chihuahuan_derived}
\alias{ushcn_chihuahuan_derived}
\title{Dataset 201 - Monthly derived climate metrics from 9 Chihuahuan desert USHCN
weather stations.}
\format{
A data frame with 11988 rows and 15 variables:
\describe{
\item{stationid}{USHCN station identifier, unique alphanumeric}
\item{date}{Date of observation, POSIXct}
\item{year}{Year of observation, integer}
\item{month}{Month of observation, integer}
\item{day}{Day of observation, integer}
\item{lat}{Latitude of station, decimal degrees}
\item{lon}{Longitude of station, decimal degrees}
\item{elev}{Elevation of station, meters}
\item{state}{State station is in, US state code string}
\item{station_name}{Unique name of station, string}
\item{tavg}{Average monthly air temperature, Celsius}
\item{ppt}{Monthly precipitation sum, millimeters}
\item{pet_tho}{Average monthly Thornthwaite potential evapotranspiration, millimeters}
\item{spei12mo}{12-month SPEI value, std deviation units}
\item{sc_pdsi}{Self-calibrating PDSI value, units?}
}
}
\source{
\describe{Derived from the ushcn_chihuahuan_data dataset and
derive201_ushcn_chihuahuan.R file}
}
\usage{
ushcn_chihuahuan_derived
}
\description{
A dataset containing stationids, names, location information, and time series
of met and derived climate indices (SPEI, scPDSI) for the 9 USHCN stations
located in the northern Chihuahuan Desert (U.S.A. portion).  Published to
EDI as Jornada dataset 201.
}
\keyword{datasets}
